requirements.txt
h5py==2.10.0
ipython==7.8.0
ipython-genutils==0.2.0
matplotlib==3.1.1
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==6.1.0
protobuf==3.9.1
scikit-image==0.14.1
scikit-learn==0.21.3
scipy==1.2.0
tensorboardX==1.8
torch==0.4.1
torchvision==0.2.1
